♻️ : #467 - MyPlan 컴포넌트 비즈니스 로직 분리
